se agrega serializacion de items completa y poder levantarlos desde el yml (el serializer solo soporta bullets y food por ahora)

+++ server_src/items/ItemSerializer.cpp
@@ -32,7 +32,7 @@ Item ItemSerializer::deserialize(const std::string& serializedItem){
 std::string ItemSerializer::serialize(Item& item){
 	std::string serializedItem;
 	
-	std::string type = std::to_string((int)item.getType());
+	std::string type = std::to_string(item.getType());
 	
 	std::string x = std::to_string(item.getPosition().getXCoordinate());
 	std::string y = std::to_string(item.getPosition().getYCoordinate());
